Changed behaviour to test_framework_drfh.
Added a thread to generate tasks randomly.
resourceOffers() iterate on offers sequentially and
uses the first with allocatable resources trying to fit
as many tasks as possible (given the ones in queue).
The stats counts the unused/declined/accepted offers.
Stats are also printed on file if afilepath is passed as argument.